cats_dogs_binary_classifier/requirements.txt
kaggle
torch
torchvision
pillow
sklearn